validate os version
